Add size and lang flags and fix some ui texts
